audio_awaiter/requirements.txt
asyncio==3.4.3
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
pika==1.3.2
pyaml==25.1.0
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
urllib3==2.3.0
